# api-specs/openapi.yaml
openapi: 3.1.0
info:
  contact:
    name: Fruition
  description: Fruition PoC 백엔드 API
  title: Fruition API
  version: 0.0.1
security:
- bearerAuth: []
...
paths:
...
  /api/workspaces/{workspaceId}/ai/tasks/{id}:
    get:
      operationId: status_1
      parameters:
      - in: path
        name: workspaceId
        required: true
        schema:
          type: string
      - in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: object
                additionalProperties: {}
          description: OK
      tags:
      - ai-task-cancellation-controller
  /api/workspaces/{workspaceId}/ai/tasks/{id}/cancel:
    post:
      operationId: cancel_1
      parameters:
      - in: path
        name: workspaceId
        required: true
        schema:
          type: string
      - in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: object
                additionalProperties: {}
          description: OK
      tags:
      - ai-task-cancellation-controller
...
  /api/workspaces/{workspace_id}/ai-model-settings:
    get:
      description: ingest·lint 작업에 쓰는 provider/model 설정을 반환합니다. OWNER와 MEMBER 모두 조회할 수 있습니다.
      operationId: get
      parameters:
      - example: ws_9d47a0e9a6324341b47562553b75f92a
        in: path
        name: workspace_id
        required: true
        schema:
          type: string
          default: ws_9d47a0e9a6324341b47562553b75f92a
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SettingsResponse"
          description: 조회 성공
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: 워크스페이스를 찾을 수 없음
      summary: 워크스페이스 AI 모델 설정 조회
      tags:
      - AI Models
    put:
      description: "ingest·lint에 쓸 provider/model을 바꿉니다. OWNER만 호출할 수 있고, 활성 model catalog에 있는 조합만 허용합니다."
      operationId: update
      parameters:
      - example: ws_9d47a0e9a6324341b47562553b75f92a
        in: path
        name: workspace_id
        required: true
        schema:
          type: string
          default: ws_9d47a0e9a6324341b47562553b75f92a
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsRequest"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SettingsResponse"
          description: 변경 성공
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: catalog에 없는 provider/model 조합
        "403":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: OWNER가 아님
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: 워크스페이스를 찾을 수 없음
      summary: 워크스페이스 AI 모델 설정 변경
      tags:
      - AI Models
...
  /internal/agent/tools/rollback/{id}/changes:
    post:
      operationId: changes
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Actor"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: array
                items:
                  type: integer
                  format: int64
          description: OK
      tags:
      - ai-task-cancellation-controller
  /internal/agent/tools/rollback/{id}/changes/{changeId}:
    post:
      operationId: undo
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      - in: path
        name: changeId
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Actor"
        required: true
      responses:
        "200":
          description: OK
      tags:
      - ai-task-cancellation-controller
  /internal/agent/tools/rollback/{id}/finalize-edits:
    post:
      operationId: finalizeEdits
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Actor"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: {}
          description: OK
      tags:
      - ai-task-cancellation-controller
...
components:
  schemas:
    Actor:
      type: object
      properties:
        user_id:
          type: string
          minLength: 1
        workspace_id:
          type: string
          minLength: 1
      required:
      - user_id
      - workspace_id
...
    AiModelSelectionRequest:
      type: object
      description: 설정할 provider와 model 조합. 활성 catalog에 있는 짝만 허용된다.
      properties:
        model:
          type: string
          description: 모델명
          example: gpt-5-nano
          minLength: 1
        provider:
          type: string
          description: LLM provider
          enum:
          - openai
          - gemini
          - claude
          example: openai
          minLength: 1
      required:
      - model
      - provider
    AiModelSelectionResponse:
      type: object
      description: 현재 설정된 provider와 model 조합
      properties:
        model:
          type: string
          description: 모델명
          example: gpt-5-nano
        provider:
          type: string
          description: LLM provider
          enum:
          - openai
          - gemini
          - claude
          example: openai
...
    ErrorDetail:
      type: object
      properties:
        code:
          type: string
          description: 분기 기준으로 쓰는 값. message가 아니라 이 값으로 분기한다.
          example: INVALID_REQUEST
        details:
          type: array
          description: 검증 실패(400 INVALID_REQUEST)일 때만 존재한다. 값이 없으면 키 자체가 빠지므로 null 체크가 아니라 존재 여부로 확인한다.
          items:
            $ref: "#/components/schemas/FieldError"
        message:
          type: string
          description: 사용자에게 그대로 보여도 되는 한국어 문구
          example: 요청 형식이 올바르지 않습니다.
    ErrorResponse:
      type: object
      description: 컨트롤러까지 도달한 요청이 실패했을 때 모든 API가 공유하는 에러 응답 형식. 인증 필터가 막는 401은 HttpStatusEntryPoint가 상태 코드만 내보내므로 이 본문이 없다.
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
          description: 에러 상세
...
    FieldError:
      type: object
      properties:
        field:
          type: string
          description: 문제가 된 요청 필드명
          example: email
        reason:
          type: string
          description: 그 필드의 위반 사유
          example: email은 필수입니다.
...
    SettingsRequest:
      type: object
      description: AI 모델 설정 변경 요청
      properties:
        ingest_lint:
          $ref: "#/components/schemas/AiModelSelectionRequest"
          description: ingest·lint에 쓸 provider/model 조합
      required:
      - ingest_lint
    SettingsResponse:
      type: object
      description: AI 모델 설정. 기본값은 gemini + gemini-3.1-flash-lite다.
      properties:
        can_update:
          type: boolean
          description: 호출자가 이 설정을 변경할 수 있는지(워크스페이스 OWNER 여부)
          example: true
        ingest_lint:
          $ref: "#/components/schemas/AiModelSelectionResponse"
          description: ingest·lint에 쓰는 provider/model 조합
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      name: bearerAuth
      scheme: bearer
      type: http